Started in on th combustion engine
I still left a good bit for you to do Arch but i wanted to make it a bit
clear how i want it to work. I use to have a combustion engine in fluid
mechanics for a while. It worked by creating and applying force to a
mechanical rod. This rod then translated the force to a motor that
actually created the power. I'll be re-adding the motor soon to the core
so we can create a few machines that work this way. As well we can
create compact versions that connect directly to wires but with less
power.