Worked over the gun render with a hammer
IT works and it works well. While implementing the gun renderer from my
empire engine it didn't go well. Since the laser doesn't left click
shoot the gun can't be rendered in the aimed strait. So for the laser i
setup both an aimed and un aimed state. As well when its aimed it
actually doesn't shoot to the top right of the player's render. Instead
it simple pulls back a bit. Gun actually looks nice and its a rather
good size when held. Now to just finish the inventory render, and entity
render for it.